+++ WWWPOS/SideBarControl/UserList/UserControl_UpdateForm.cs
@@ -63,7 +63,7 @@ namespace WWWPOS.SideBarControl.UserList
             {
                 DataBase DB = new DataBase();
 
-                int phoneNumber = Int32.Parse(txt_Number.Text);
+                long phoneNumber = Int64.Parse(txt_Number.Text);
                 int accountID = Int32.Parse(txt_AccountID.Text);
 
                 DB.UpdateUser(accountID, txt_Name.Text, txt_Email.Text, txt_Password.Text, phoneNumber, txt_UserType.Text, txt_Address.Text);
